VEES-18 | Refactor `InfoBlock`
