+++ README.md
@@ -32,7 +32,7 @@
 
 ⠀
 
-### 🫡 Общие правила проекта
+### 📋✏️ Общие правила проекта
 
 #### Архитектура
 
@@ -62,7 +62,9 @@
 
 ---
 
-### 👾 Скрипты
+⠀
+
+### 🎬📃 Скрипты
 
 - `yarn dev` - Запуск в Development режиме;
 - `yarn build` - Production сборка;
@@ -71,7 +73,9 @@
 
 ---
 
-### ⚡️ Архитектура проекта
+⠀
+
+### 📂⚡️ Архитектура проекта
 
 Проект написан в соответствии с архитектурной методологией [Feature-Sliced Design](https://feature-sliced.design/docs/get-started/tutorial) и [AppRouter](https://nextjs.org/docs/app) Next.js 15.
