+++ docs/naming-conventions.md
@@ -1,0 +1,4 @@
+- Vees — Тренировка
+- All Vees — Все тренировки
+- Exercise — Упражнение
+- Muscle — Группа мышц
