+++ src/app/(with-navbar)/page.tsx
@@ -1,9 +1,10 @@
+import { Card } from '@/widgets/Card';
+import { mockExerciseListItems } from '@/widgets/Card/model/mock';
 import { Headers } from '@/widgets/Headers';
-import { InfoBlock } from '@/widgets/InfoBlock/client';
-import { mockInfoBlockExercisesItems } from '@/widgets/InfoBlock/model/mock-exercises';
 
-import { ArrowLeftIcon, FilterOutlineIcon } from '@/shared/icons';
-import { Badge, Button, Flex, Layout, Separator } from '@/shared/ui';
+import { ArrowLeftIcon, FilterOutlineIcon, StopwatchOutlineIcon } from '@/shared/icons';
+import { Badge, Button, Flex, Layout, List, Separator } from '@/shared/ui';
+import { InfoBlock } from '@/shared/ui/client';
 
 import { FiltersModal } from '@/features/FiltersModal';
 import { SortModal } from '@/features/SortModal';
@@ -15,11 +16,36 @@ const Home = () => (
     <Headers.Main />
     <main>
       <Layout>
-        <InfoBlock
-          items={mockInfoBlockExercisesItems}
-          showAllCountItems={2}
-          showAllText="Показать всю статистику"
-        />
+        <InfoBlock maxHeight={126} textButton="Показать всю статистику">
+          <List.Item
+            description="~ конец в 18:36"
+            icon={<StopwatchOutlineIcon />}
+            title="Расчётное время"
+            value="~ 1 час 30 минут"
+          />
+          <List.Horizontal
+            items={[
+              {
+                title: 'Тренировка №',
+                value: '98',
+              },
+              {
+                title: 'Начало',
+                value: '17:06',
+              },
+              {
+                title: 'Упражнений',
+                value: '7',
+              },
+            ]}
+          />
+          <List.Item
+            icon={<StopwatchOutlineIcon />}
+            showDivider={false}
+            title="Пусто"
+            value="Пусто"
+          />
+        </InfoBlock>
         <Separator size={18} />
         <Flex gap={10} id="home-filters-buttons">
           <FiltersModal>
@@ -51,6 +77,10 @@ const Home = () => (
             </Button>
           </SortModal>
         </Flex>
+      </Layout>
+      <Separator size={32} />
+      <Layout padding={false}>
+        <Card.ExercisesList items={mockExerciseListItems} />
       </Layout>
     </main>
   </>
