+++ src/shared/ui/InfoBlock/client.ts
@@ -1,0 +1,1 @@
+export { InfoBlock } from './InfoBlock';
