+++ src/shared/ui/List/ui/Horizontal/horizontal.module.scss
@@ -1,0 +1,4 @@
+.wrapper {
+  padding-top: 6px;
+  padding-bottom: 6px;
+}
