+++ src/shared/ui/client.ts
@@ -1,2 +1,3 @@
 export * from './Image/client';
+export * from './InfoBlock/client';
 export * from './Modal/client';
