+++ src/widgets/Card/ui/Card/card.module.scss
@@ -1,0 +1,39 @@
+.wrapper {
+  display: flex;
+
+  animation: var(--anim-scale-fade-in);
+
+  @media (max-width: theme('screens.sm')) {
+    border-radius: .75rem 0 0 .75rem;
+  }
+}
+
+.content {
+  position: relative;
+  padding: 15px 10px 15px 0;
+}
+
+.chipList {
+  position: absolute;
+  top: -12px;
+
+  color: theme('colors.base.600');
+}
+
+.buttonShow {
+  width: 64px;
+  height: auto;
+
+  &.collapsed {
+    color: theme('colors.base.600');
+    background: theme('colors.base.800');
+
+    &:hover {
+      background: theme('colors.base.700') !important;
+    }
+  }
+
+  @media (max-width: theme('screens.sm')) {
+    border-radius: .75rem 0 0 .75rem;
+  }
+}
