+++ src/widgets/Card/ui/VeesItem/veesItem.module.scss
@@ -1,43 +1,3 @@
-.wrapper {
-  display: flex;
-
-  animation: var(--anim-scale-fade-in);
-
-  @media (max-width: theme('screens.sm')) {
-    border-radius: .75rem 0 0 .75rem;
-  }
-}
-
-.content {
-  position: relative;
-  padding: 15px 10px 15px 0;
-}
-
-.chipList {
-  position: absolute;
-  top: -12px;
-
-  color: theme('colors.base.600');
-}
-
-.buttonShow {
-  width: 64px;
-  height: auto;
-
-  &.collapsed {
-    color: theme('colors.base.600');
-    background: theme('colors.base.800');
-
-    &:hover {
-      background: theme('colors.base.700') !important;
-    }
-  }
-
-  @media (max-width: theme('screens.sm')) {
-    border-radius: .75rem 0 0 .75rem;
-  }
-}
-
 .button {
   gap: 2;
   min-width: 80px;
