+++ src/widgets/Card/ui/VeesList/VeesList.tsx
@@ -1,14 +1,12 @@
-'use client';
-
 import { FinishOutlineIcon } from '@/shared/icons';
 import { Button, Flex } from '@/shared/ui';
 
-import { IVeesItem, VeesItem } from '../VeesItem/VeesItem';
+import { ICardVeesItem, VeesItem } from '../VeesItem/VeesItem';
 
 import styles from './veesList.module.scss';
 
 interface IVeesList {
-  items: IVeesItem[]
+  items: ICardVeesItem[]
 }
 
 export const VeesList = ({ items }: IVeesList) => (
@@ -16,6 +14,7 @@ export const VeesList = ({ items }: IVeesList) => (
     {items.map((item, index) => (
       <VeesItem item={item} key={`${index}${item.name}`} />
     ))}
+    {/* TODO: Вынести кнопку в features */}
     <Button
       className={styles.button}
       icon={<FinishOutlineIcon height={22} width={22} />}
