+++ src/widgets/Card/ui/VeesList/veesList.module.scss
