--- src/widgets/InfoBlock/client.ts
@@ -1,1 +1,0 @@
-export { InfoBlock } from './ui/InfoBlock';
