--- src/widgets/InfoBlock/index.ts
@@ -1,1 +1,0 @@
-export type { IInfoBlockItem } from './ui/InfoBlock';
