--- src/widgets/Navbar/ui/Navbar/navbar.global.scss
@@ -1,3 +1,0 @@
-main > *:last-child {
-  padding-bottom: 160px;
-}
