--- src/widgets/VeesList/client.ts
@@ -1,1 +1,0 @@
-export { VeesList } from './ui/VeesList/VeesList';
